Move `RecordSource` to `recordsource` package.
