+++ service/src/main/java/org/databiosphere/workspacedataservice/recordsource/JsonRecordSource.java
@@ -1,7 +1,5 @@
 package org.databiosphere.workspacedataservice.recordsource;
 
-import static org.databiosphere.workspacedataservice.service.BatchWriteService.*;
-
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonToken;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -9,7 +7,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.List;
-import org.databiosphere.workspacedataservice.service.BatchWriteService.WriteStreamInfo;
 import org.databiosphere.workspacedataservice.shared.model.BatchOperation;
 import org.databiosphere.workspacedataservice.shared.model.OperationType;
 import org.databiosphere.workspacedataservice.shared.model.Record;
