+++ service/src/main/java/org/databiosphere/workspacedataservice/recordsource/RecordSourceFactory.java
@@ -9,7 +9,6 @@ import org.apache.avro.file.DataFileStream;
 import org.apache.avro.generic.GenericRecord;
 import org.apache.parquet.hadoop.ParquetReader;
 import org.databiosphere.workspacedataservice.recordsource.TwoPassRecordSource.ImportMode;
-import org.databiosphere.workspacedataservice.service.BatchWriteService.RecordSource;
 import org.databiosphere.workspacedataservice.service.model.TdrManifestImportTable;
 import org.databiosphere.workspacedataservice.shared.model.RecordType;
 import org.springframework.stereotype.Component;
