+++ service/src/main/java/org/databiosphere/workspacedataservice/recordsource/TwoPassRecordSource.java
@@ -1,7 +1,5 @@
 package org.databiosphere.workspacedataservice.recordsource;
 
-import org.databiosphere.workspacedataservice.service.BatchWriteService.RecordSource;
-
 /**
  * Marker interface that denotes a RecordSource which runs in two passes. The first pass will upsert
  * base attributes, and the second pass will upsert relation attributes. This allows for data
